Fix validator import and rename course to courses
