+++ packages/api/src/app.module.ts
@@ -6,7 +6,7 @@ import { join } from 'path';
 import { env } from '@/common/env';
 
 import { AuthModule } from './auth/auth.module';
-import { CourseModule } from './courses/course.module';
+import { CoursesModule } from './courses/courses.module';
 import { LearningMaterialsModule } from './learning-materials/learning-materials.module';
 import { PrismaModule } from './prisma/prisma.module';
 import { UsersModule } from './users/users.module';
@@ -25,10 +25,10 @@ const productionImports = [
     ...(isProduction ? productionImports : []),
     PrismaModule,
     UsersModule,
-    CourseModule,
     AuthModule,
     LearningMaterialsModule,
     CqrsModule,
+    CoursesModule,
   ],
 })
 export class AppModule {}
