+++ packages/shared/package.json
@@ -6,6 +6,7 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
+    "class-transformer": "0.4.0",
     "class-validator": "0.13.1"
   }
 }
